feat: add oauth, github, google, etc...

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma"
 import * as yup from 'yup';
+import { getUserSessionServer } from "@/auth";
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url)
@@ -14,9 +15,17 @@ export async function GET(request: Request) {
     }, { status: 400 });
   }
 
+  const user = await getUserSessionServer()
+  if (!user) {
+    return NextResponse.json({
+      message: 'Not authorized'
+    }, { status: 401 })
+  }
+
   const todos = await prisma.todo.findMany({
     skip: (page - 1) * limit, // de donde comienza
     take: limit, // cuantos
+    where: { userId: user.id }
   });
 
   return NextResponse.json(todos);
@@ -28,11 +37,14 @@ const CreateTodoValidation = yup.object({
 })
 
 export async function POST(request: Request) {
-  const body = await request.json();
+  const body = await request.json()
+
+  const user = await getUserSessionServer()
+  if (!user) return NextResponse.json({ message: 'Not authorized' }, { status: 401 })
 
   try {
     const { completed, description } = await CreateTodoValidation.validate(body);
-    const todo = await prisma.todo.create({ data: { completed, description } })
+    const todo = await prisma.todo.create({ data: { completed, description, userId: user.id } })
     return NextResponse.json(todo);
   } catch (error: any) {
     if (error.name === 'ValidationError') {
@@ -43,8 +55,11 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServer()
+  if (!user) return NextResponse.json({ message: 'Not authorized' }, { status: 401 })
+
   try {
-    await prisma.todo.deleteMany({ where: { completed: true } })
+    await prisma.todo.deleteMany({ where: { completed: true, userId: user.id } })
     return NextResponse.json({ message: 'Deleted completed todos' });
   } catch (error: any) {
     if (error.name === 'ValidationError') {
